test: split compact command lifecycle coverage
